requirements.txt
Markdown==3.10.2
numpy==2.4.2
pandas==3.0.1
python-dateutil==2.9.0.post0
python-frontmatter==1.1.0
PyYAML==6.0.3
six==1.17.0
